Add custom SQL parser and improve Calcite integration (#957)
* Add schemas for all formats

* Add sql parser files copied from calcite

* Add ddl directive to support the WITH clause in CREATE TABLE statements

+++ baremaps-data/src/main/java/org/apache/baremaps/data/memory/MemoryMappedDirectory.java
@@ -116,7 +116,9 @@ public class MemoryMappedDirectory extends Memory<MappedByteBuffer> {
     header.clear();
     segments.clear();
 
-    // Delete the directory and all files in it
-    FileUtils.deleteRecursively(directory);
+    // Delete the directory and all files in it if it exists
+    if (Files.exists(directory)) {
+      FileUtils.deleteRecursively(directory);
+    }
   }
 }
